+++ src/physical_calculator/physical_calculator.hpp
@@ -35,7 +35,7 @@ public:
   //! \brief it adds a box in the world
   btRigidBody* addBox( btVector3 size, btVector3 position, btScalar mass);
   //! it calculates the next step
-  void nextStep(float time=1/80.f, int addedoperations=20);
+  void nextStep(double time=1/80.f, int addedoperations=20);
 
   unsigned long int getTime() const;
   //void addRobotToScene( Robot * robot, Wheel * md, Wheel *mg, Wheel *ed, Wheel *eg);
